requirements.txt
colorama==0.4.6
defusedxml==0.7.1
fonttools==4.61.1
fpdf2==2.8.5
pillow==12.0.0